% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Permutations_With_Repetition}
\alias{Permutations_With_Repetition}
\title{Enumerate the Permutations of the Elements of a Vector When Some of those Elements are Identical}
\usage{
Permutations_With_Repetition(Sample_Sizes)
}
\arguments{
\item{Sample_Sizes}{Numeric vector (\code{n1},...,\code{nk}) that indicates the number of times each element is repeated.}
}
\value{
Returns a matrix where each row contains a permutation.
}
\description{
This function enumerates the possible combinations of \code{n} elements where the first element is repeated \code{n1} times, the sencond element is repeated \code{n2} times, the third \code{n3} times, ...
}
\section{Warning}{

The number of permutations increases rapidly with lenght (Sets). The computational time increases exponetianly with the number of elements and with the number of sets.
}

\examples{
Sample_Sizes <- c(2,2,2)
Permutations_With_Repetition(Sample_Sizes)

Sample_Sizes <- c(2,3)
Permutations_With_Repetition(Sample_Sizes)

Permutations_With_Repetition(c(3,2,3))

Permutations_With_Repetition(c(10,5))
}
